Add goal to implement context system
The context system is basically a way to to print out messages from a source to a client. There has been an implementation that I've thought of:

Make the Source model have an argument called context which will be given by the server. There will also be a stub implementation of context so that individual sources can be used on their own without any server giving some context to it, which would be the default value.